Fix crash at startup
send_scheduler() seemed to get called very early when there are
no subtunnels there. In that case the % operator failed because then
we do %0.

This commit fixes this by calculating the size of $subtun_sessions
every run. (Maybe not so performant). And explicitely catching the cases
where $subtun_count is 0 and doing nothing then